+++ src/app/admin/permissions/RoleView.tsx
@@ -5,7 +5,7 @@ import CreateRoleForm from './CreateRoleForm'
 import DeleteRoleForm from './DeleteRoleForm'
 import { UpdateRoleForm } from './UpdateRoleForm'
 import UserManagmentForm from './UserManagmentForm'
-import { readUsersOfRole } from '@/actions/permissions/read'
+import { readUsersOfRole } from '@/actions/rolePermissions/read'
 import React, { useEffect, useState } from 'react'
 import { v4 as uuid } from 'uuid'
 import Link from 'next/link'
