+++ src/auth/authoptions.ts
@@ -3,7 +3,7 @@ import PrismaAdapter from './feide/PrismaAdapter'
 import signUp from './feide/signUp'
 import { updateFeideAccount } from './feide/index'
 import prisma from '@/prisma'
-import { readPermissionsOfUser } from '@/actions/permissions/read'
+import { readPermissionsOfUser } from '@/actions/rolePermissions/read'
 import CredentialsProvider from 'next-auth/providers/credentials'
 import { decode } from 'next-auth/jwt'
 import type { JWT } from 'next-auth/jwt'
